requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
fake-useragent==1.4.0
scikit-learn==1.3.0
numpy==1.24.3
werkzeug==2.3.7
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-mail==0.9.1
flask-bcrypt==1.0.1
itsdangerous==2.1.2
firebase-admin==6.4.0
pyrebase4==4.7.1

